+++ src/esm/components/navigation/header-nav/components/burger-menu.js
@@ -2,7 +2,7 @@ import _objectWithoutProperties from "@babel/runtime/helpers/esm/objectWithoutPr
 import _extends from "@babel/runtime/helpers/esm/extends";
 import React, { useRef } from 'react';
 import { BurgerIcon } from '../../../../components/icons/burger-icon';
-import { Dropdown } from '../../../../components/dropdowns/v2/dropdown';
+import { Dropdown } from './dropdown';
 import { VisuallyHidden } from '../../../../components/accessibility/visually-hidden';
 import COLORS from '../../../../constants/colors-config';
 import { Context } from './context';